requirements.txt
python-telegram-bot==21.6
anthropic==0.40.0
chromadb==0.5.23
python-dotenv==1.0.1
openai>=1.0.0
